Aktualizacja 19.04.2023

+++ CarFuelEconomyApp/ICarType.cs
@@ -7,9 +7,9 @@ namespace CarFuelEconomyApp
         public string Brand { get; set; }
         public string Plate { get; set; }
         public void AddGrade (float grade);
-       
-        public void AddGrade(decimal grade);
-        public  void AddGrade(string grade);
+
+        public void AddGrade(int grade);
+        public void AddGrade(string grade);
         
         Statistics GetStatistics();
     }
